requirements.txt
# PrimateReID dependencies
# ML framework
torch>=2.0
torchvision>=0.15
# Embedding
facenet-pytorch>=2.5
# Image processing
Pillow>=10.0
opencv-python>=4.8
# Scientific computing
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
# Detection (future)
ultralytics>=8.0
# Config
pyyaml>=6.0
# Utilities
tqdm>=4.65
